refactor: relocate some components to atomic folders
