+++ frontend/src/features/about/components/organisms/MiniGameBoard.tsx
